third_party/WebKit/LayoutTests/fast/compositorworker/compositor-proxy-disconnect-worker.html

<!DOCTYPE html>
<div id="container"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

var test = async_test(function() {
    function processMessage(event) {
        var message = event.data;
        assert_equals('NoModificationAllowedError', message);
        test.done();
    }

    var worker = new CompositorWorker('resources/proxy-disconnect.js');
    worker.onmessage = processMessage;
    var proxy = new CompositorProxy(document.getElementById('container'), ['opacity']);
    worker.postMessage(proxy);
}, "This test checks that an element's compositor proxy cannot be mutated after disconnect().");
</script>
